+++ bluelight_api/src/main/java/com/c5r/bluelight_api/QuestionLike/QuestionLike.java
@@ -21,19 +21,19 @@ public class QuestionLike {
             strategy = GenerationType.SEQUENCE,
             generator = "question_like_id_seq"
     )
-    private Integer id;
+    private Long id;
 
     @Column(
             name = "USER_ID",
             nullable = false
     )
-    private Integer userId;
+    private Long userId;
 
     @Column(
             name = "QUESTION_ID",
             nullable = false
     )
-    private Integer questionId;
+    private Long questionId;
 
     @Column(
             name = "WAS_LIKED",
@@ -49,15 +49,15 @@ public class QuestionLike {
     )
     private boolean wasDisliked = false;
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public void setQuestionId(Integer questionId) {
+    public void setQuestionId(Long questionId) {
         this.questionId = questionId;
     }
 
@@ -69,15 +69,15 @@ public class QuestionLike {
         this.wasDisliked = wasDisliked;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return this.userId;
     }
 
-    public Integer getQuestionId() {
+    public Long getQuestionId() {
         return this.questionId;
     }
 
@@ -91,7 +91,7 @@ public class QuestionLike {
 
     public QuestionLike() {}
 
-    public QuestionLike(Integer userId, Integer questionId, boolean wasLiked, boolean wasDisliked) {
+    public QuestionLike(long userId, long questionId, boolean wasLiked, boolean wasDisliked) {
         this.userId = userId;
         this.questionId = questionId;
         this.wasLiked = wasLiked;
