+++ bluelight_api/src/main/java/com/c5r/bluelight_api/UserQuestion/AnswerRequest.java
@@ -2,14 +2,14 @@ package com.c5r.bluelight_api.UserQuestion;
 
 public class AnswerRequest {
     private String userId;
-    private int questionId;
+    private Long questionId;
     private boolean isCorrect;
 
     public String getUserId() {
         return this.userId;
     }
 
-    public int getQuestionId() {
+    public Long getQuestionId() {
         return this.questionId;
     }
